+++ handlers/add_email.py
@@ -1,0 +1,15 @@
+from decorators.input_error import input_error
+from models.adressbook import AddressBook
+
+@input_error
+def add_email(args, contacts: AddressBook):
+    if len(args) != 2:
+        raise ValueError
+    name, email = args
+
+    if name in contacts:
+        record = contacts[name]
+        record.add_email(email)
+        return f"Email added for {name}."
+    else:
+        raise KeyError
